+++ CHANGELOG.md
@@ -4,7 +4,11 @@ Release notes for the OOPSpam Craft CMS plugin.
 
 ## 1.3.0- 2025-09-09
 ### Added
+- Added `Craft Commerce` integration
 - Added contextual spam detection
+
+### Changed
+- Improved `verbb/comments` integration
 
 ## 1.2.1 - 2025-06-16
 ### Added